stress: test_get_device_list: change printf format for ssize_t
The type of list_size is ssize_t.
First idea was to use "%zd" printf format specifier.
But that is not supported by Visual Studio 2013.

Use "%ld" + cast list_size to long.

Suggested-by: Chris Dickens <christopher.a.dickens@gmail.com>

+++ tests/stress.c
@@ -60,8 +60,8 @@ static libusb_testlib_result test_get_device_list(libusb_testlib_ctx * tctx)
 		ssize_t list_size = libusb_get_device_list(ctx, &device_list);
 		if (list_size < 0 || device_list == NULL) {
 			libusb_testlib_logf(tctx,
-				"Failed to get device list on iteration %d: %d (%p)",
-				i, -list_size, device_list);
+				"Failed to get device list on iteration %d: %ld (%p)",
+				i, (long)-list_size, device_list);
 			return TEST_STATUS_FAILURE;
 		}
 		libusb_free_device_list(device_list, 1);
